requirements.txt
streamlit
pandas
scikit-learn
numpy
pydantic
pyyaml